add bnum_integer initial version
This is an alternative implementation for large integer arithmetic.
It is based on 'bnum' crate, which is supposed to behave like Rust's
primitive integer types. Integers are stored as an array of digits
(primitive unsigned integer) of fixed length, which allows to keep them
on the stack since they are fixed size.
That should result of much better performance.

Types defined:
- signed: BnumI256, BnumI512.
- usigned: BnumU256, BnumU512.