Add handling for cases when one of the resources could be zero.
This commit adds handling and tests to cases where if one of the resources has a low divisibility we could get into cases where we calculate that we need a zero contribution of the low divisibility resource and X of the higher divisibility one. Which is not the correct handling of this case. If this happens, it is now a hard error.