Fix tests.
The tests failed since the `hashset!` and `hashmap!` macros defined in
the `radix-rust` crate did not use the correct hasher according to the
enabled features. This lead the macro to expand to something like
`HashSet<T, StubHasher>::new()` and the function `new` is not defined
for `HashSet`s that don't use the default `RandomState` hasher.